prototype/backend/requirements.txt
alembic==1.0.10
aniso8601==7.0.0
asn1crypto==0.24.0
attrs==19.1.0
bcrypt==3.1.6
cffi==1.12.3
Click==7.0
cryptography==2.7
Flask==1.0.3
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-HTTPAuth==3.3.0
Flask-Migrate==2.5.2
flask-restplus==0.12.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-Testing==0.7.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
Mako==1.0.12
MarkupSafe==1.1.1
pycparser==2.19
PyJWT==1.7.1
pyrsistent==0.15.2
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.5
Werkzeug==0.15.4